+++ app/sprinkles/core/tests/Integration/Database/Migrator/MigrationLocatorTest.php
@@ -90,6 +90,7 @@ class MigrationLocatorTest extends TestCase
             new Resource($resourceStream, $resourceAccountLocation, 'one/CreatePasswordResetsTable.php'),
             new Resource($resourceStream, $resourceAccountLocation, 'two/CreateFlightsTable.php'),
             new Resource($resourceStream, $resourceAccountLocation, 'CreateMainTable.php'),
+            new Resource($resourceStream, $resourceAccountLocation, 'README.md'), // This shoudn't be returned by the migrator
         ]);
 
         // Create a new MigrationLocator instance with our simulated ResourceLocation
@@ -109,7 +110,6 @@ class MigrationLocatorTest extends TestCase
         ];
 
         // Test results match expectations
-        $this->assertCount(8, $results);
         $this->assertEquals($expected, $results);
 
         return $locator;